requirements.txt
plotly==4.14.3
streamlit==0.76.0
pandas==1.1.3
lime==0.2.0.1
lightgbm==3.1.1
numpy==1.19.2
matplotlib==3.3.2
xgboost==1.3.3
seaborn==0.11.0
gTTS==2.2.2
catboost==0.24.4
imbalanced_learn==0.7.0
mlxtend==0.18.0
imblearn==0.0
Pillow==8.1.0
scikit_learn==0.24.1
